+++ vehicle/src/Vehicle/Backend/LossFunction/Logics.hs
@@ -20,6 +20,12 @@ import GHC.Generics (Generic)
 import Vehicle.Backend.Prelude (DifferentiableLogic (..))
 import Vehicle.Compile.Prelude qualified as V
 
+-- Target flag --target json
+-- Resolve instance arguments
+-- Erase all implicit arguments
+-- (Later) Make relevant implicit arguments explicit
+-- Move jsonOptions to somewhere, and add ToJSON, FromJSON.
+
 jsonOptions :: Options
 jsonOptions =
   defaultOptions
@@ -115,12 +121,12 @@ data DifferentialLogicImplementation = DifferentialLogicImplementation
 
 implementationOf :: DifferentiableLogic -> DifferentialLogicImplementation
 implementationOf = \case
-  DL2 -> dl2Translation
-  Godel -> godelTranslation
-  Lukasiewicz -> lukasiewiczTranslation
-  Product -> productTranslation
-  Yager -> yagerTranslation
-  STL -> stlTranslation
+  DL2Loss -> dl2Translation
+  GodelLoss -> godelTranslation
+  LukasiewiczLoss -> lukasiewiczTranslation
+  ProductLoss -> productTranslation
+  YagerLoss -> yagerTranslation
+  STLLoss -> stlTranslation
 
 --------------------------------------------------------------------------------
 -- different available  differentiable logics