+++ vehicle/src/Vehicle/Expr/Normalisable.hs
@@ -1,5 +1,6 @@
 module Vehicle.Expr.Normalisable where
 
+import Data.Aeson (ToJSON)
 import Data.Hashable (Hashable)
 import Data.Serialize
 import GHC.Generics
@@ -14,7 +15,7 @@ data NormalisableBuiltin types
   = CConstructor BuiltinConstructor
   | CFunction BuiltinFunction
   | CType types
-  deriving (Eq, Show, Generic)
+  deriving (Eq, Ord, Show, Generic)
 
 instance (Pretty types) => Pretty (NormalisableBuiltin types) where
   pretty = \case
@@ -25,6 +26,8 @@ instance (Pretty types) => Pretty (NormalisableBuiltin types) where
 instance (Serialize types) => Serialize (NormalisableBuiltin types)
 
 instance (Hashable types) => Hashable (NormalisableBuiltin types)
+
+instance (ToJSON types) => ToJSON (NormalisableBuiltin types)
 
 -----------------------------------------------------------------------------
 -- Expressions