+++ vehicle/src/Vehicle/Verify/Specification/IO.hs
@@ -1,9 +1,10 @@
 module Vehicle.Verify.Specification.IO
   ( readSpecification,
     readVerificationPlan,
-    outputVerificationResult,
+    outputCompilationResults,
     verifySpecification,
     verificationPlanFileName,
+    isValidQueryFolder,
   )
 where
 
@@ -15,13 +16,14 @@ import Data.Aeson (decode)
 import Data.Aeson.Encode.Pretty (encodePretty')
 import Data.ByteString.Lazy qualified as BIO
 import Data.List.NonEmpty (NonEmpty (..))
-import Data.Map qualified as Map
-import Data.Text (unpack)
+import Data.Text (intercalate, pack, unpack)
 import Data.Text.Encoding (decodeUtf8)
 import Data.Text.IO qualified as TIO
 import Data.Text.Lazy qualified as Text
-import System.Directory (createDirectoryIfMissing)
+import System.Directory (createDirectoryIfMissing, doesFileExist)
+import System.Exit (exitFailure)
 import System.FilePath (takeExtension, (<.>), (</>))
+import System.IO (stderr, stdout)
 import System.ProgressBar
 import Vehicle.Backend.Prelude
 import Vehicle.Compile.Prelude
@@ -64,19 +66,22 @@ readSpecification inputFile
 -- | Outputs the compiled verification plan and queries that make up the specification.
 -- If a folder is provided it outputs them to individual files in that folder,
 -- otherwise it outputs them to stdout.
-outputVerificationResult ::
-  (MonadIO m, MonadLogger m) =>
+outputCompilationResults ::
+  (MonadVerify m) =>
   QueryFormatID ->
   Maybe FilePath ->
   (VerificationPlan, VerificationQueries) ->
   m ()
-outputVerificationResult queryFormatID maybeOutputLocation (plan, queries) = do
+outputCompilationResults queryFormatID maybeOutputLocation (plan, queries) = do
   case maybeOutputLocation of
     Nothing -> return ()
     Just folder -> liftIO $ createDirectoryIfMissing True folder
 
   outputVerificationPlan maybeOutputLocation plan
   writeVerificationQueries queryFormatID maybeOutputLocation queries
+
+isValidQueryFolder :: (MonadIO m) => FilePath -> m Bool
+isValidQueryFolder folder = liftIO $ doesFileExist (verificationPlanFileName folder)
 
 outputVerificationPlan :: (MonadIO m) => Maybe FilePath -> VerificationPlan -> m ()
 outputVerificationPlan maybeFolder plan = do
@@ -126,7 +131,7 @@ readVerificationPlan planFile = do
       Just plan -> return plan
 
 writeVerificationQueries ::
-  (MonadIO m, MonadLogger m) =>
+  (MonadVerify m) =>
   QueryFormatID ->
   Maybe FilePath ->
   VerificationQueries ->
@@ -150,50 +155,61 @@ verificationPlanFileName folder = folder </> "" <.> vehicleVerificationPlanFileE
 --------------------------------------------------------------------------------
 -- Verification
 
+type MonadVerify m =
+  ( MonadLogger m,
+    MonadIO m
+  )
+
+type MonadVerifyProperty m =
+  ( MonadVerify m,
+    MonadReader (Verifier, VerifierExecutable, FilePath, PropertyProgressBar) m
+  )
+
 -- | Uses the verifier to verify the specification. Failure of one property does
 -- not prevent the verification of the other properties.
 verifySpecification ::
-  (MonadIO m) =>
+  (MonadVerify m) =>
   FilePath ->
   Verifier ->
   VerifierExecutable ->
-  Specification QueryMetaData ->
+  Specification (Property QueryMetaData) ->
+  Maybe FilePath ->
   m SpecificationStatus
-verifySpecification queryFolder verifier verifierExecutable (Specification namedProperties) = do
+verifySpecification queryFolder verifier verifierExecutable (Specification namedProperties) assignmentsLocation = do
   programOutput "Verifying properties:"
   results <- forM namedProperties $ \(name, property) -> do
-    result <- verifyMultiProperty verifier verifierExecutable queryFolder property
+    result <- verifyMultiProperty verifier verifierExecutable queryFolder assignmentsLocation property
     return (name, result)
-  return $ SpecificationStatus (Map.fromList results)
+
+  return $ Specification results
 
 verifyMultiProperty ::
   forall m.
-  (MonadIO m) =>
+  (MonadVerify m) =>
   Verifier ->
   VerifierExecutable ->
   FilePath ->
-  MultiProperty QueryMetaData ->
+  Maybe FilePath ->
+  MultiProperty (Property QueryMetaData) ->
   m MultiPropertyStatus
-verifyMultiProperty verifier verifierExecutable queryFolder = go
+verifyMultiProperty verifier verifierExecutable queryFolder assignmentsLocation = go
   where
-    go :: MultiProperty QueryMetaData -> m MultiPropertyStatus
+    go :: MultiProperty (Property QueryMetaData) -> m MultiPropertyStatus
     go = \case
-      MultiProperty ps -> MultiPropertyStatus <$> traverse go ps
+      MultiProperty ps -> MultiProperty <$> traverse go ps
       SingleProperty address property -> do
         progressBar <- createPropertyProgressBar address (propertySize property)
         let readerState = (verifier, verifierExecutable, queryFolder, progressBar)
         result <- runReaderT (verifyProperty property) readerState
-        return $ SinglePropertyStatus result
-
-type MonadVerify m =
-  ( MonadReader (Verifier, VerifierExecutable, FilePath, PropertyProgressBar) m,
-    MonadIO m
-  )
+        liftIO $ TIO.putStrLn (layoutAsText $ "    result: " <> pretty result)
+        outputAssignments assignmentsLocation result
+        return $ SingleProperty address result
 
 -- | Lazily tries to verify the property, avoiding evaluating parts
 -- of the expression that are not needed.
 verifyProperty ::
-  (MonadVerify m) =>
+  forall m.
+  (MonadVerifyProperty m) =>
   Property QueryMetaData ->
   m PropertyStatus
 verifyProperty property = do
@@ -201,9 +217,8 @@ verifyProperty property = do
   return $ PropertyStatus negationStatus status
   where
     go ::
-      (MonadVerify m) =>
       BooleanExpr (QuerySet QueryMetaData) ->
-      m (QueryNegationStatus, MaybeTrivial (QueryResult UserVariableCounterexample))
+      m (QuerySetNegationStatus, MaybeTrivial (QueryResult UserVariableAssignment))
     go = \case
       Query qs -> verifyQuerySet qs
       Disjunct x y -> do
@@ -218,9 +233,9 @@ verifyProperty property = do
           else go y
 
 verifyQuerySet ::
-  (MonadVerify m) =>
+  (MonadVerifyProperty m) =>
   QuerySet QueryMetaData ->
-  m (QueryNegationStatus, MaybeTrivial (QueryResult UserVariableCounterexample))
+  m (QuerySetNegationStatus, MaybeTrivial (QueryResult UserVariableAssignment))
 verifyQuerySet (QuerySet negated queries) = case queries of
   Trivial b -> return (negated, Trivial b)
   NonTrivial disjuncts -> do
@@ -229,15 +244,14 @@ verifyQuerySet (QuerySet negated queries) = case queries of
 
 verifyDisjunctAll ::
   forall m.
-  (MonadVerify m) =>
+  (MonadVerifyProperty m) =>
   DisjunctAll (QueryAddress, QueryMetaData) ->
-  m (QueryResult UserVariableCounterexample)
+  m (QueryResult UserVariableAssignment)
 verifyDisjunctAll (DisjunctAll ys) = go ys
   where
     go ::
-      (Monad m) =>
       NonEmpty (QueryAddress, QueryMetaData) ->
-      m (QueryResult UserVariableCounterexample)
+      m (QueryResult UserVariableAssignment)
     go (x :| []) = verifyQuery x
     go (x :| y : xs) = do
       r <- verifyQuery x
@@ -246,24 +260,56 @@ verifyDisjunctAll (DisjunctAll ys) = go ys
         else go (y :| xs)
 
 verifyQuery ::
-  (MonadVerify m) =>
+  (MonadVerifyProperty m) =>
   (QueryAddress, QueryMetaData) ->
-  m (QueryResult UserVariableCounterexample)
+  m (QueryResult UserVariableAssignment)
 verifyQuery (queryAddress, QueryData metaNetwork userVar) = do
   (verifier, verifierExecutable, folder, progressBar) <- ask
   let queryFile = folder </> calculateQueryFileName queryAddress
-  result <- invokeVerifier verifier verifierExecutable metaNetwork queryFile
-  liftIO $ incProgress progressBar 1
-  return $ fmap (reconstructUserVars userVar) result
+  errorOrResult <- invokeVerifier verifier verifierExecutable metaNetwork queryFile
+  case errorOrResult of
+    Left errMsg -> liftIO $ do
+      TIO.hPutStrLn stderr ("\nError: " <> errMsg)
+      exitFailure
+    Right result -> do
+      liftIO $ incProgress progressBar 1
+      traverse (reconstructUserVars metaNetwork userVar) result
+
+--------------------------------------------------------------------------------
+-- Assignments
+
+outputAssignments ::
+  (MonadIO m) =>
+  Maybe FilePath ->
+  PropertyStatus ->
+  m ()
+outputAssignments maybeLocation (PropertyStatus _negated s) = case s of
+  NonTrivial (SAT (Just (UserVariableAssignment assignments))) ->
+    case maybeLocation of
+      Nothing -> do
+        let assignmentDocs = vsep (fmap prettyUserVariableAssignment assignments)
+        let witnessDoc = indent 6 assignmentDocs
+        liftIO $ TIO.hPutStrLn stdout (layoutAsText witnessDoc)
+      Just _location -> liftIO $ do
+        fatalError "outputting assignments to a file not yet implemented"
+  {-
+  let folder = takeDirectory location </> layoutAsString (pretty address)
+  createDirectoryIfMissing True folder
+  forM_ assignments $ \UserVariableAssignment {..} -> do
+    let file = folder </> Text.unpack variableName
+    let idxData = _
+    encodeIDXFile _ file
+  -}
+  _ -> return ()
 
 --------------------------------------------------------------------------------
 -- Calculation of file paths
 
 calculateQueryFileName :: QueryAddress -> FilePath
-calculateQueryFileName ((propertyName, propertyIndices), queryID) = do
+calculateQueryFileName (PropertyAddress propertyName propertyIndices, queryID) = do
   let propertyStr
         | null propertyIndices = ""
-        | otherwise = concatMap (\v -> "!" <> show v) (reverse propertyIndices)
+        | otherwise = showTensorIndices propertyIndices
 
   unpack propertyName
     <> propertyStr
@@ -273,15 +319,16 @@ calculateQueryFileName ((propertyName, propertyIndices), queryID) = do
 type PropertyProgressBar = ProgressBar ()
 
 createPropertyProgressBar :: (MonadIO m) => PropertyAddress -> Int -> m PropertyProgressBar
-createPropertyProgressBar (name, _indices) numberOfQueries = do
+createPropertyProgressBar (PropertyAddress name indices) numberOfQueries = do
+  let propertyName = Text.fromStrict $ intercalate "!" (name : fmap (pack . show) indices)
   let style =
         defStyle
-          { stylePrefix = msg ("  " <> Text.fromStrict name),
-            stylePostfix = exact <> msg " queries complete",
+          { stylePrefix = msg ("  " <> propertyName),
+            stylePostfix = exact <> msg " queries",
             styleWidth = ConstantWidth 80
           }
   let initialProgress = Progress 0 numberOfQueries ()
-  liftIO $ newProgressBar style 10 initialProgress
+  liftIO $ hNewProgressBar stdout style 10 initialProgress
 
 {-
 completeProgress :: MonadIO m => PropertyProgressBar -> m ()