+++ vehicle/src/Vehicle/Verify/Specification/Status.hs
@@ -1,20 +1,22 @@
 module Vehicle.Verify.Specification.Status where
 
+import Control.Monad (forM_)
 import Data.Aeson
 import Data.List.Split (chunksOf)
-import Data.Map (Map)
-import Data.Map qualified as Map
 import Data.Text (Text, pack)
 import Data.Vector.Unboxed qualified as Vector
 import GHC.Generics (Generic)
 import System.Console.ANSI (Color (..))
 import Vehicle.Compile.Prelude
 import Vehicle.Compile.Print (prettyFriendly)
-import Vehicle.Compile.Type.Subsystem.Standard (TypeCheckedExpr)
+import Vehicle.Compile.Queries.Variable
+import Vehicle.Compile.Type.Subsystem.Standard.Core
 import Vehicle.Compile.Type.Subsystem.Standard.Patterns
 import Vehicle.Expr.Boolean (MaybeTrivial (..))
+import Vehicle.Expr.DeBruijn (Ix)
 import Vehicle.Expr.Normalisable (NormalisableBuiltin (..))
 import Vehicle.Verify.Core
+import Vehicle.Verify.Specification
 
 class IsVerified a where
   isVerified :: a -> Bool
@@ -24,7 +26,7 @@ instance IsVerified (QueryResult witness) where
     SAT {} -> True
     UnSAT -> False
 
-evaluateQuery :: QueryNegationStatus -> MaybeTrivial (QueryResult witness) -> Bool
+evaluateQuery :: QuerySetNegationStatus -> MaybeTrivial (QueryResult witness) -> Bool
 evaluateQuery negated q =
   negated `xor` case q of
     Trivial b -> b
@@ -34,67 +36,69 @@ evaluateQuery negated q =
 -- Verification status of a single property
 
 data PropertyStatus
-  = PropertyStatus QueryNegationStatus (MaybeTrivial (QueryResult UserVariableCounterexample))
+  = PropertyStatus QuerySetNegationStatus (MaybeTrivial (QueryResult UserVariableAssignment))
   deriving (Generic)
 
 instance FromJSON PropertyStatus
 
 instance ToJSON PropertyStatus
 
+instance IsVerified PropertyStatus where
+  isVerified (PropertyStatus negated result) = evaluateQuery negated result
+
+instance Pretty PropertyStatus where
+  pretty (PropertyStatus negated s) = do
+    let witnessText = if negated then "counterexample" else "witness"
+    let (verified, evidenceText) = case s of
+          Trivial status -> (status `xor` negated, "(trivial)")
+          NonTrivial status -> case status of
+            UnSAT -> (negated, "proved no" <+> witnessText <+> "exists")
+            SAT Nothing -> (not negated, "no" <> witnessText <+> "found")
+            SAT Just {} -> (not negated, witnessText <+> "found")
+    pretty (statusSymbol verified) <+> "-" <+> evidenceText
+
 --------------------------------------------------------------------------------
 -- Verification status of a multi property
 
-data MultiPropertyStatus
-  = MultiPropertyStatus [MultiPropertyStatus]
-  | SinglePropertyStatus PropertyStatus
-  deriving (Generic)
-
-instance FromJSON MultiPropertyStatus
-
-instance ToJSON MultiPropertyStatus
+type MultiPropertyStatus = MultiProperty PropertyStatus
 
 instance IsVerified MultiPropertyStatus where
   isVerified = \case
-    MultiPropertyStatus ps -> and (fmap isVerified ps)
-    SinglePropertyStatus (PropertyStatus negated result) -> evaluateQuery negated result
+    MultiProperty ps -> and (fmap isVerified ps)
+    SingleProperty _ status -> isVerified status
 
-prettyPropertyStatus :: Name -> MultiPropertyStatus -> Doc a
-prettyPropertyStatus name = \case
-  MultiPropertyStatus ps -> do
-    let numberedSubproperties =
-          zipWith (\(i :: Int) p -> (name <> "!" <> pack (show i), p)) [0 ..] ps
+nameSubProperties :: Name -> [MultiPropertyStatus] -> [(Name, MultiPropertyStatus)]
+nameSubProperties name = zipWith (\(i :: Int) p -> (name <> "!" <> pack (show i), p)) [0 ..]
+
+prettyMultiPropertyStatus :: Name -> MultiPropertyStatus -> Doc a
+prettyMultiPropertyStatus name = \case
+  MultiProperty ps -> do
+    let namedSubproperties = nameSubProperties name ps
     let numVerified = pretty (length (filter isVerified ps))
     let num = pretty $ length ps
-    let summary = pretty name <> ":" <+> numVerified <> "/" <> num <+> "verified"
-    let results = indent 2 $ vsep (fmap (uncurry prettyPropertyStatus) numberedSubproperties)
+    let summary = "Property" <+> quotePretty name <> ":" <+> numVerified <> "/" <> num <+> "verified"
+    let results = indent 2 $ vsep (fmap (uncurry prettyMultiPropertyStatus) namedSubproperties)
     summary <> line <> results
-  SinglePropertyStatus (PropertyStatus negated s) -> do
-    let witnessText = if negated then "Counter-example" else "Witness"
-    let (verified, evidenceText) = case s of
-          Trivial status -> (status `xor` negated, " (trivial)")
-          NonTrivial status -> case status of
-            UnSAT -> (negated, "")
-            SAT Nothing -> (not negated, witnessText <> ": none")
-            SAT (Just witness) -> do
-              let assignments = vsep (fmap prettyUserVariableAssignment witness)
-              let witnessDoc = witnessText <> line <> indent 2 assignments
-              (not negated, witnessDoc)
+  SingleProperty _address status ->
+    pretty name <+> pretty status
 
-    prettyNameAndStatus name verified <> line <> evidenceText
+statusSymbol :: Bool -> String
+statusSymbol verified = do
+  let (colour, symbol) = if verified then (Green, "🗸") else (Red, "✗")
+  setTextColour colour symbol
 
 prettyNameAndStatus :: Text -> Bool -> Doc a
 prettyNameAndStatus name verified = do
-  let (colour, symbol) = if verified then (Green, "🗸") else (Red, "✗")
-  pretty (setTextColour colour symbol) <+> pretty name
+  pretty (statusSymbol verified) <+> pretty name
 
-prettyUserVariableAssignment :: UserVariableAssignment -> Doc a
-prettyUserVariableAssignment UserVariableAssignment {..} = do
-  let name = pretty variableName
-  let valueExpr = assignmentToExpr variableDimensions (Vector.toList variableValue)
+prettyUserVariableAssignment :: (UserVariable, VariableValue) -> Doc a
+prettyUserVariableAssignment (UserVariable {..}, variableValue) = do
+  let name = pretty userVarName
+  let valueExpr = assignmentToExpr userVarDimensions (Vector.toList variableValue)
   let value = prettyFriendly (WithContext valueExpr emptyDBCtx)
   name <> ":" <+> value
 
-assignmentToExpr :: TensorDimensions -> [Double] -> TypeCheckedExpr
+assignmentToExpr :: TensorDimensions -> [Double] -> Expr Ix StandardBuiltin
 assignmentToExpr [] xs = RatLiteral mempty (toRational (head xs))
 assignmentToExpr (dim : dims) xs = do
   let vecConstructor = Builtin mempty (CConstructor $ LVec dim)
@@ -102,24 +106,36 @@ assignmentToExpr (dim : dims) xs = do
   let elems = fmap (ExplicitArg mempty . assignmentToExpr dims) inputVarIndicesChunks
   normAppList mempty vecConstructor elems
 
+traverseMultiPropertySATResults ::
+  forall m.
+  (Monad m) =>
+  (PropertyAddress -> QuerySetNegationStatus -> Maybe UserVariableAssignment -> m ()) ->
+  MultiPropertyStatus ->
+  m ()
+traverseMultiPropertySATResults f = go
+  where
+    go :: MultiPropertyStatus -> m ()
+    go = \case
+      MultiProperty ps -> do
+        forM_ ps go
+      SingleProperty address (PropertyStatus negated p) -> case p of
+        Trivial {} -> return ()
+        NonTrivial UnSAT -> return ()
+        NonTrivial (SAT assignment) -> do
+          f address negated assignment
+
 --------------------------------------------------------------------------------
 -- Verification status of the specification
 
-newtype SpecificationStatus = SpecificationStatus (Map Name MultiPropertyStatus)
-  deriving (Generic)
-
-instance FromJSON SpecificationStatus
-
-instance ToJSON SpecificationStatus
+type SpecificationStatus = Specification PropertyStatus
 
 instance IsVerified SpecificationStatus where
-  isVerified (SpecificationStatus properties) =
-    and (fmap isVerified (Map.elems properties))
+  isVerified (Specification properties) =
+    and (fmap (isVerified . snd) properties)
 
 instance Pretty SpecificationStatus where
-  pretty spec@(SpecificationStatus properties) = do
-    let result = "Result:" <+> (if isVerified spec then "true" else "false")
-    let propertiesByName = Map.toList properties
+  pretty spec@(Specification properties) = do
+    let result = "Specification summary:" <+> (if isVerified spec then "true" else "false")
     result
       <> line
-      <> indent 2 (vsep (fmap (uncurry prettyPropertyStatus) propertiesByName))
+      <> indent 2 (vsep (fmap (uncurry prettyMultiPropertyStatus) properties))