+++ vehicle/src/Vehicle/Verify/Verifier/Marabou.hs
@@ -8,7 +8,6 @@ import Data.List (elemIndex, findIndex)
 import Data.Text (Text)
 import Data.Text qualified as Text (pack, splitOn, strip, unpack)
 import Data.Text.IO (hPutStrLn)
-import Data.Vector.Unboxed (Vector)
 import Data.Vector.Unboxed qualified as Vector
 import System.Exit (ExitCode (..), exitFailure)
 import System.IO (stderr)
@@ -25,7 +24,7 @@ marabouVerifier =
     { verifierIdentifier = Marabou,
       verifierExecutableName = "Marabou",
       invokeVerifier = invokeMarabou,
-      verifierQueryFormat = MarabouQueryFormat
+      verifierQueryFormat = MarabouQueries
     }
 
 --------------------------------------------------------------------------------
@@ -41,9 +40,9 @@ invokeMarabou marabouExecutable networkLocations queryFile =
     parseMarabouOutput command marabouOutput
 
 prepareNetworkArg :: MetaNetwork -> IO String
-prepareNetworkArg [(_name, file)] = return file
+prepareNetworkArg [MetaNetworkEntry {..}] = return metaNetworkEntryFilePath
 prepareNetworkArg metaNetwork = do
-  let duplicateNetworkNames = findDuplicates (fmap fst metaNetwork)
+  let duplicateNetworkNames = findDuplicates (fmap metaNetworkEntryName metaNetwork)
 
   let errorMsg =
         "Error: Marabou currently doesn't support properties that involve"
@@ -51,7 +50,7 @@ prepareNetworkArg metaNetwork = do
             then
               "multiple networks. This property involves:"
                 <> line
-                <> indent 2 (vsep $ fmap (\(n, _) -> "the network" <+> squotes (pretty n)) metaNetwork)
+                <> indent 2 (vsep $ fmap (\e -> "the network" <+> squotes (pretty $ metaNetworkEntryName e)) metaNetwork)
             else
               "multiple applications of the same network. This property applies:"
                 <> line
@@ -60,34 +59,48 @@ prepareNetworkArg metaNetwork = do
   hPutStrLn stderr $ layoutAsText errorMsg
   exitFailure
 
-parseMarabouOutput :: String -> (ExitCode, String, String) -> IO (QueryResult NetworkVariableCounterexample)
+parseMarabouOutput ::
+  String ->
+  (ExitCode, String, String) ->
+  IO (Either Text (QueryResult NetworkVariableAssignment))
 parseMarabouOutput command (exitCode, out, _err) = case exitCode of
-  ExitFailure _ -> do
-    -- Marabou seems to output its error messages to stdout rather than stderr...
-    let errorDoc =
-          "Marabou threw the following error:"
-            <> line
-            <> indent 2 (pretty out)
-            <> "when running the command:"
-            <> line
-            <> indent 2 (pretty command)
-    hPutStrLn stderr (layoutAsText errorDoc)
-    exitFailure
+  ExitFailure exitValue
+    | exitValue < 0 -> do
+        -- Marabou was killed by the system.
+        -- See System.Process.html#waitForProcess documentation
+        let errorDoc =
+              "Marabou was automatically killed by the OS with signal"
+                <+> quotePretty (-exitValue)
+                <> "."
+                <> line
+                <> "The most common reason for this is an out-of-memory-error."
+        return $ Left $ layoutAsText errorDoc
+    | otherwise -> do
+        -- Marabou seems to output its error messages to stdout rather than stderr...
+        let errorDoc =
+              "Marabou threw the following error:"
+                <> line
+                <> indent 2 (pretty out)
+                <> line
+                <> "when running the command:"
+                <> line
+                <> indent 2 (pretty command)
+        hPutStrLn stderr (layoutAsText errorDoc)
+        exitFailure
   ExitSuccess -> do
-    -- print (layoutAsString $ pretty (lines out))
     let outputLines = fmap Text.pack (lines out)
     let resultIndex = findIndex (\v -> v == "sat" || v == "unsat") outputLines
     case resultIndex of
       Nothing -> malformedOutputError "cannot find 'sat' or 'unsat'"
       Just i
         | outputLines !! i == "unsat" ->
-            return UnSAT
+            return $ Right UnSAT
         | otherwise -> do
             let assignmentOutput = drop (i + 1) outputLines
             ioVarAssignment <- parseSATAssignment (filter (/= "") assignmentOutput)
-            return $ SAT $ Just ioVarAssignment
+            return $ Right $ SAT $ Just ioVarAssignment
 
-parseSATAssignment :: [Text] -> IO (Vector Double)
+parseSATAssignment :: [Text] -> IO NetworkVariableAssignment
 parseSATAssignment output = do
   let mInputIndex = elemIndex "Input assignment:" output
   let mOutputIndex = elemIndex "Output:" output
@@ -97,7 +110,7 @@ parseSATAssignment output = do
       let outputVarLines = drop (outputIndex + 1) output
       let inputValues = parseSATAssignmentLine Input <$> inputVarLines
       let outputValues = parseSATAssignmentLine Output <$> outputVarLines
-      return $ Vector.fromList (inputValues <> outputValues)
+      return $ NetworkVariableAssignment $ Vector.fromList (inputValues <> outputValues)
     _ -> malformedOutputError "could not find strings 'Input assignment:' and 'Output:'"
 
 parseSATAssignmentLine :: InputOrOutput -> Text -> Double