+++ vehicle/tests/golden/Vehicle/Test/Golden/TestSpec/NewTestSpec.hs
@@ -39,6 +39,7 @@ import System.FilePath
     isRelative,
     normalise,
     takeDirectory,
+    takeExtension,
     (</>),
   )
 import Test.Tasty (Timeout)
@@ -57,33 +58,24 @@ import Vehicle.Compile qualified as CompileOptions
     target,
   )
 import Vehicle.Compile qualified as Vehicle (CompileOptions)
-import Vehicle.CompileAndVerify qualified as CompileAndVerifyOptions
-  ( datasetLocations,
-    networkLocations,
-    proofCache,
-    specification,
-    verifierID,
-  )
-import Vehicle.CompileAndVerify qualified as Vehicle (CompileAndVerifyOptions)
 import Vehicle.Export qualified as ExportOptions
   ( outputFile,
     proofCacheLocation,
     target,
   )
 import Vehicle.Export qualified as Vehicle (ExportOptions)
-import Vehicle.Prelude (Pretty (pretty), layoutAsString)
+import Vehicle.Prelude (Pretty (pretty), layoutAsString, vehicleSpecificationFileExtension)
 import Vehicle.Test.Golden.Extra (createDirectoryRecursive)
 import Vehicle.Test.Golden.TestSpec
-  ( FilePattern,
-    TestSpec (..),
+  ( TestSpec (..),
     TestSpecs (TestSpecs),
     addOrReplaceTestSpec,
     encodeTestSpecsPretty,
-    filePatternString,
-    parseFilePattern,
     readTestSpecsFile,
     writeTestSpecsFile,
   )
+import Vehicle.Test.Golden.TestSpec.FilePattern (GoldenFilePattern)
+import Vehicle.Test.Golden.TestSpec.FilePattern qualified as FilePattern
 import Vehicle.TypeCheck qualified as TypeCheckOptions
   ( specification,
   )
@@ -96,7 +88,7 @@ import Vehicle.Verify qualified as VerifyOptions
     proofCache,
     verifierID,
   )
-import Vehicle.Verify.Core (QueryFormatID (MarabouQueryFormat))
+import Vehicle.Verify.Core (QueryFormatID (MarabouQueries))
 
 data NewTestSpecOptions = NewTestSpecOptions
   { newTestSpecDryRun :: Bool,
@@ -162,10 +154,9 @@ newTestSpec args = do
       fail $
         printf "Test needs files at an absolute path: %s\n" testSpecNeed
   forM_ testSpecProduces $ \testSpecProducePattern ->
-    let testSpecProduce = filePatternString testSpecProducePattern
-     in unless (isRelative testSpecProduce) $
-          fail $
-            printf "Test produces files at an absolute path: %s\n" testSpecProduce
+    unless (FilePattern.isRelative testSpecProducePattern) $
+      fail $
+        printf "Test produces files at an absolute path: %s\n" (show testSpecProducePattern)
 
   -- Construct the test specification:
   let theNewTestSpec =
@@ -175,8 +166,9 @@ newTestSpec args = do
             testSpecEnabled = Nothing,
             testSpecNeeds = testSpecDataNeeds,
             testSpecProduces = testSpecProduces,
+            testSpecExternal = [],
             testSpecTimeout = newTestSpecTestTimeout,
-            testSpecDiffSpec = Nothing
+            testSpecIgnore = Nothing
           }
 
   -- Write the test:
@@ -217,7 +209,7 @@ newTestSpec args = do
 data TestSpecData = TestSpecData
   { testSpecDataTarget :: String,
     testSpecDataNeeds :: [FilePath],
-    testSpecDataProduces :: Either String [FilePattern]
+    testSpecDataProduces :: Either String [GoldenFilePattern]
   }
   deriving (Show)
 
@@ -228,7 +220,7 @@ class TestSpecLike a where
   needs :: a -> [FilePath]
   needs = testSpecDataNeeds . testSpecData
 
-  produces :: a -> Either String [FilePattern]
+  produces :: a -> Either String [GoldenFilePattern]
   produces = testSpecDataProduces . testSpecData
 
   testSpecData :: a -> TestSpecData
@@ -249,7 +241,6 @@ instance TestSpecLike Vehicle.ModeOptions where
   testSpecData = \case
     ModeOptions.Check opts -> testSpecData opts
     ModeOptions.Compile opts -> testSpecData opts
-    ModeOptions.CompileAndVerify opts -> testSpecData opts
     ModeOptions.Verify opts -> testSpecData opts
     ModeOptions.Export opts -> testSpecData opts
     ModeOptions.Validate opts -> testSpecData opts
@@ -263,7 +254,7 @@ instance TestSpecLike Vehicle.TypeCheckOptions where
     [ TypeCheckOptions.specification opts
     ]
 
-  produces :: Vehicle.TypeCheckOptions -> Either String [FilePattern]
+  produces :: Vehicle.TypeCheckOptions -> Either String [GoldenFilePattern]
   produces = const (return [])
 
 instance TestSpecLike Vehicle.CompileOptions where
@@ -278,13 +269,13 @@ instance TestSpecLike Vehicle.CompileOptions where
         Map.elems (CompileOptions.datasetLocations opts)
       ]
 
-  produces :: Vehicle.CompileOptions -> Either String [FilePattern]
-  produces opts = traverse parseFilePattern filePatternStrings
+  produces :: Vehicle.CompileOptions -> Either String [GoldenFilePattern]
+  produces opts = traverse FilePattern.readEither filePatternStrings
     where
       outputFile = CompileOptions.outputFile opts
       filePatternStrings =
         case CompileOptions.target opts of
-          VerifierQueries MarabouQueryFormat ->
+          VerifierQueries MarabouQueries ->
             [outputDir </> "*.txt" | outputDir <- maybeToList outputFile]
           _ -> maybeToList outputFile
 
@@ -293,29 +284,21 @@ instance TestSpecLike Vehicle.VerifyOptions where
   targetName = layoutAsString . pretty . VerifyOptions.verifierID
 
   needs :: Vehicle.VerifyOptions -> [FilePath]
-  needs opts = do
-    join
-      [ [VerifyOptions.queryFolder opts]
-      -- TODO the verification plan also references resources and query files
-      ]
+  needs opts
+    | takeExtension (VerifyOptions.specification opts) == vehicleSpecificationFileExtension = do
+        join
+          [ [VerifyOptions.specification opts]
+          -- TODO the verification plan also references resources and query files
+          ]
+    | otherwise =
+        join
+          [ [VerifyOptions.specification opts],
+            Map.elems (VerifyOptions.networkLocations opts),
+            Map.elems (VerifyOptions.datasetLocations opts)
+          ]
 
-  produces :: Vehicle.VerifyOptions -> Either String [FilePattern]
-  produces = traverse parseFilePattern . maybeToList . VerifyOptions.proofCache
-
-instance TestSpecLike Vehicle.CompileAndVerifyOptions where
-  targetName :: Vehicle.CompileAndVerifyOptions -> String
-  targetName = layoutAsString . pretty . CompileAndVerifyOptions.verifierID
-
-  needs :: Vehicle.CompileAndVerifyOptions -> [FilePath]
-  needs opts =
-    join
-      [ [CompileAndVerifyOptions.specification opts],
-        Map.elems (CompileAndVerifyOptions.networkLocations opts),
-        Map.elems (CompileAndVerifyOptions.datasetLocations opts)
-      ]
-
-  produces :: Vehicle.CompileAndVerifyOptions -> Either String [FilePattern]
-  produces = traverse parseFilePattern . maybeToList . CompileAndVerifyOptions.proofCache
+  produces :: Vehicle.VerifyOptions -> Either String [GoldenFilePattern]
+  produces = traverse FilePattern.readEither . maybeToList . VerifyOptions.proofCache
 
 instance TestSpecLike Vehicle.ExportOptions where
   targetName :: Vehicle.ExportOptions -> String
@@ -324,8 +307,8 @@ instance TestSpecLike Vehicle.ExportOptions where
   needs :: Vehicle.ExportOptions -> [FilePath]
   needs = (: []) . ExportOptions.proofCacheLocation
 
-  produces :: Vehicle.ExportOptions -> Either String [FilePattern]
-  produces = traverse parseFilePattern . maybeToList . ExportOptions.outputFile
+  produces :: Vehicle.ExportOptions -> Either String [GoldenFilePattern]
+  produces = traverse FilePattern.readEither . maybeToList . ExportOptions.outputFile
 
 instance TestSpecLike Vehicle.ValidateOptions where
   targetName :: Vehicle.ValidateOptions -> String
@@ -334,5 +317,5 @@ instance TestSpecLike Vehicle.ValidateOptions where
   needs :: Vehicle.ValidateOptions -> [FilePath]
   needs = (: []) . ValidateOptions.proofCache
 
-  produces :: Vehicle.ValidateOptions -> Either String [FilePattern]
+  produces :: Vehicle.ValidateOptions -> Either String [GoldenFilePattern]
   produces = const (return [])